shoes__shop/shoe_shop-main/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-extensions==4.1
mysqlclient==2.2.8
pillow==12.1.1
pygraphviz==1.14
sqlparse==0.5.5
